BAU: Extract `LogEventMatcher` to `shared-test`
